mnma-chat/requirements.txt
langchain
langchain-core
langfuse
langchain-qdrant
langchain_aws
langchain_community
boto3
fastapi-cognito
qdrant-client
aiohttp
async-timeout
asyncio
grpcio
httpcore
httpx
nest-asyncio
PyYAML
requests
fastapi
pydantic
python-multipart
uvicorn[standard]
SQLAlchemy
pytz
textstat
numpy
aiofiles>=23.0.0